first refactor attempt, moving send_domain_invitation_email to a helper
