+++ src/registrar/models/domain.py
@@ -7,6 +7,7 @@ from typing import Optional
 from django_fsm import FSMField, transition, TransitionNotAllowed  # type: ignore
 
 from django.db import models
+from django.urls import reverse
 from django.utils import timezone
 from typing import Any
 from registrar.models.host import Host
@@ -1086,6 +1087,13 @@ class Domain(TimeStampedModel, DomainHelper):
         blank=True,
         help_text="Record of the last change event for ds data",
     )
+
+    def get_absolute_url(self):
+        """
+        Returns the absolute URL for the domain instance.
+        This is the standard implementation for Django models.
+        """
+        return reverse("domain", kwargs={"pk": self.pk})
 
     def isActive(self):
         return self.state == Domain.State.CREATED
