+++ src/registrar/utility/errors.py
@@ -23,6 +23,34 @@ class InvalidDomainError(ValueError):
     pass
 
 
+class InvitationError(Exception):
+    """Base exception for invitation-related errors."""
+
+    pass
+
+
+class AlreadyManagerError(InvitationError):
+    """Raised when the user is already a manager for the domain."""
+
+    def __init__(self, email):
+        super().__init__(f"{email} is already a manager for this domain.")
+
+
+class AlreadyInvitedError(InvitationError):
+    """Raised when the user has already been invited to the domain."""
+
+    def __init__(self, email):
+        super().__init__(f"{email} has already been invited to this domain.")
+
+
+class MissingEmailError(InvitationError):
+    """Raised when the requestor has no email associated with their account."""
+
+    def __init__(self, username):
+        super().__init__(f"Can't send invitation email. No email is associated with the account for '{username}'.")
+        self.username = username
+
+
 class OutsideOrgMemberError(ValueError):
     """
     Error raised when an org member tries adding a user from a different .gov org.
